backend/requirements.txt
Flask>=3.0.0
flask-cors>=4.0.0
flask-jwt-extended>=4.6.0
Flask-SQLAlchemy>=3.1.1
python-dotenv>=1.0.1
Werkzeug>=3.0.0
